requirements.txt
pyserial
schedule
